% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.r
\docType{data}
\name{fire}
\alias{fire}
\title{Gives distance and damage of a fire}
\format{
A data frame with 15 rows and 2 variables
\describe{
\item{DISTANCE}{distance unit}
\item{DAMAGE}{damage unit}
}
}
\usage{
fire
}
\description{
This dataset contains the distance a fire traveled and the amount of damage the fire did.
}
\keyword{datasets}
