% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753chen0122.R
\docType{package}
\name{MATH4753chen0122}
\alias{MATH4753chen0122}
\title{MATH4753 package}
\description{
Assortment of functions used in and taken from OU MATH4753
}
\author{
Eric Chen \email{eric.h.chen-1@ou.edu}
}
