% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Create confidence interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample}

\item{fun}{function to be used}

\item{alpha}{alpha value for confidence interval}

\item{cx}{graph modifier}
}
\value{
invisible list of vectors/lists in function and a histogram with interval
}
\description{
This function, given a sample set, applies a specified function to the set, and creates a confidence interval based on a specified alpha value. It also creates a histogram of the distribution and plots the interval.
}
\examples{
myboot2(x=sam, alpha=0.05, iter=10000, fun = "mean")
myboot2(x=sam, alpha=0.3, iter=10000, fun = "sd")

}
