% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Single Sample Confidence Interval}
\usage{
myci(x, conf.level = 0.95)
}
\arguments{
\item{x}{sample}

\item{conf.level}{confidence level}
}
\value{
interval contained in a list
}
\description{
This function takes in a sample and confidence level to create a confidence interval
}
