% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{myclt}
\alias{myclt}
\title{Frequency of Sample Sums}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{size of samples}

\item{iter}{number of iterations}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
Histogram and sums of samples
}
\description{
This function takes in a lower and upper limit, as well as sample size and iterations, and generates a histogram based on the frequency of the sums of these samples.
}
\examples{
myclt(n=10, iter=10000)

}
