% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimate}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function}

\item{x}{sample list}

\item{param}{parameters}
}
\description{
This functions takes in a function, sample list, and their respective parameters to find the maximum likelihood estimate
}
\examples{
\dontrun{mymaxlik(lfun=logbin, x=y, param=seq(0, 1, length=1000))}
\dontrun{mymaxlik(lfun=logpoiss, x = y, param = seq(0, 10, length = 1000))}

}
