% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Creates, displays, and calculates the area/probability of a given region in a normal distribution plot}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{upper limit (probability P(X<=a))}
}
\value{
Displays the curve with the shaded region along with the area, and outputs the area as a list in the command line
}
\description{
This function takes in the mean, standard deviation, and upper limit (of the area) to calculate and display the area under the curve.
}
\examples{
myncurve(2,5,5)
myncurve(5,1,5)

}
