% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Find P-value and plots tcalc with rejection regions}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{tcalc value}

\item{xmax}{plot bounds}

\item{n}{sample size}

\item{alpha}{alpha value}
}
\value{
gives a quantile value and the p-value. Also plots the distribution with the rejection regions shaded in, the p-value as the title, and the tcalc values marked.
}
\description{
This function, given the tcalc value, will find the p-value, and plot the distribution with the rejection regions shaded in green. The function also marks the locations of the tcalc values and lets the user determine whether a null hypothesis can be rejected or not
}
\examples{
mypvalue(tcalc, n=30, alpha=0.05)
mypvalue(tcalc, n=20, alpha=0.1)

}
