% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{Reads a CSV file}
\usage{
myread(csv)
}
\arguments{
\item{Path}{to CSV file}
}
\value{
Dataframe type
}
\description{
This function takes in a csv, reads through its contents,
and creates a dataframe by seperating the terms by commas.
}
\examples{
mpg.df=myread("EPAGAS.csv")

}
