% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{Reads a CSV file}
\usage{
myread(csv)
}
\arguments{
\item{csv}{Name of csv file}
}
\value{
Dataframe type
}
\description{
This function takes in a csv, reads through its contents,
and creates a dataframe by separating the terms by commas.
If the file is not in the current directory, users will be prompted
to locate and choose the file off their system.
}
\examples{
\dontrun{mpg.df=myread("EPAGAS.csv")}
\dontrun{choose.df=myread("")}

}
